detectService/requirements.txt
aio-pika
roboflow
pydantic
python-dotenv
requests